# uiuc_incas_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: INCAS TA2-UIUC Datatypes
  description: This API document is defined based on INCAS Common Datatypes version
    0.0.3.
  contact:
    name: Shengzhong Liu (sl29@illinois.edu)
  version: 1.0.0
servers:
- url: https://incas.cs.illinois.edu:8443/api/v1
  description: UIUC INCAS API Server
paths:
...
  /actor/{id}:
    get:
      tags:
      - actor
      description: Returns specific actor by id
      operationId: actor_id_get
      parameters:
      - name: id
        in: path
        description: Actor ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: A list of actor enrichments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
        "404":
          description: Not found
        "400":
          description: Invalid query
      x-openapi-router-controller: uiuc_incas_server.controllers.actor_controller
  /actor/enrichments:
    get:
      tags:
      - actor
      description: "Returns current actor enrichment meta by providerName, enrichmentName\
        \ and version"
      operationId: actor_enrichments_get
      parameters:
      - name: enrichmentName
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - category
          - numerical
          - array
          - text
      - name: providerName
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: version
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: A list of actor enrichment meta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorEnrichmentMeta'
      x-openapi-router-controller: uiuc_incas_server.controllers.actor_controller
    put:
      tags:
      - actor
      description: "Updates actor enrichment meta (after all actors have been added)\
        \ by providerName, enrichmentName and version"
      operationId: actor_enrichments_put
      parameters:
      - name: enrichmentName
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - category
          - numerical
          - array
          - text
      - name: providerName
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: version
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        description: The new enrichment meta to update
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ActorEnrichmentMeta'
        required: true
      responses:
        "200":
          description: A list of newly added actor enrichment meta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorEnrichmentMeta'
        "400":
          description: Invalid input
        "409":
          description: Input already exists
      x-openapi-router-controller: uiuc_incas_server.controllers.actor_controller
    post:
      tags:
      - actor
      description: Creates actor enrichment meta (post after all actors have been
        added)
      operationId: actor_enrichments_post
      requestBody:
        description: The new enrichment meta to add
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ActorEnrichmentMeta'
        required: true
      responses:
        "201":
          description: A list of newly added Actor enrichment meta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorEnrichmentMeta'
        "400":
          description: Invalid input
        "409":
          description: Input already exists
      x-openapi-router-controller: uiuc_incas_server.controllers.actor_controller
    delete:
      tags:
      - message
      description: "Delete specific message enrichment meta by providerName, enrichmentName\
        \ and version"
      operationId: actor_enrichments_delete
      parameters:
      - name: enrichmentName
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - category
          - numerical
          - array
          - text
      - name: providerName
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: version
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: id not found
        "401":
          description: Unauthorized
      x-openapi-router-controller: uiuc_incas_server.controllers.message_controller
  /actor/{id}/enrichments:
    get:
      tags:
      - actor
      description: "Returns all matched enrichment for the specific actor by type,\
        \ providerName and version"
      operationId: actor_id_enrichments_get
      parameters:
      - name: id
        in: path
        description: Actor ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: enrichmentName
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - category
          - numerical
          - array
          - text
      - name: providerName
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: version
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: A list of actor enrichments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActorEnrichment'
                x-content-type: application/json
        "404":
          description: Not found
        "400":
          description: Invalid query
      x-openapi-router-controller: uiuc_incas_server.controllers.actor_controller
    put:
      tags:
      - actor
      description: "Update the enrichments for specific actor by type, providerName\
        \ and version"
      operationId: actor_id_enrichments_put
      parameters:
      - name: id
        in: path
        description: Message ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: enrichmentName
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
          enum:
          - category
          - numerical
          - array
          - text
      - name: providerName
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        description: The new enrichments to update
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ActorEnrichment'
        required: true
      responses:
        "200":
          description: A list of updated actor enrichments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActorEnrichment'
                x-content-type: application/json
        "404":
          description: Not found
        "400":
          description: Invalid query
      x-openapi-router-controller: uiuc_incas_server.controllers.actor_controller
    post:
      tags:
      - actor
      description: Creates new enrichments for specific message
      operationId: actor_id_enrichments_post
      parameters:
      - name: id
        in: path
        description: Actor ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: The new enrichment to add
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ActorEnrichment'
        required: true
      responses:
        "201":
          description: A list of new actor enrichments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActorEnrichment'
                x-content-type: application/json
        "400":
          description: Invalid input
        "409":
          description: Input already exists
      x-openapi-router-controller: uiuc_incas_server.controllers.actor_controller
    delete:
      tags:
      - actor
      description: "Delete the enrichments for specific actor by type, providerName\
        \ and version"
      operationId: actor_id_enrichments_delete
      parameters:
      - name: id
        in: path
        description: Actor ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: enrichmentName
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
          enum:
          - category
          - numerical
          - array
          - text
      - name: providerName
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: id not found
        "401":
          description: Unauthorized
      x-openapi-router-controller: uiuc_incas_server.controllers.actor_controller
...
components:
  schemas:
...
    BaseActorEnrichmentMeta:
      required:
      - enrichmentType
      type: object
      properties:
        providerName:
          type: string
          description: "The team (e.g., UIUC-DMG) who provides the enrichment."
        enrichmentName:
          type: string
          description: "The enrichment (e.g., Concern-Stance) name for the enrichment."
        version:
          type: string
          description: "The version within the same (provider, name)."
        enrichmentType:
          type: string
          description: For discriminator
          enum:
          - category
          - numerical
          - array
          - text
      description: Stores the meta data related to enrichment algorithm.
      discriminator:
        propertyName: enrichmentType
        mapping:
          category: '#/components/schemas/CategoryActorEnrichmentMeta'
          numerical: '#/components/schemas/NumericalActorEnrichmentMeta'
          array: '#/components/schemas/ArrayActorEnrichmentMeta'
          text: '#/components/schemas/TextActorEnrichmentMeta'
...
    CategoryActorEnrichmentMeta:
      allOf:
      - $ref: '#/components/schemas/BaseActorEnrichmentMeta'
      - type: object
        additionalProperties:
          type: string
        description: Defines all available categories in category actor enrichment.
    NumericalActorEnrichmentMeta:
      allOf:
      - $ref: '#/components/schemas/BaseActorEnrichmentMeta'
      - type: object
    ArrayActorEnrichmentMeta:
      allOf:
      - $ref: '#/components/schemas/BaseActorEnrichmentMeta'
      - type: object
        additionalProperties:
          type: string
        description: "Defines the hash map of (posID, class) for multi-class regression/classification."
    TextActorEnrichmentMeta:
      allOf:
      - $ref: '#/components/schemas/BaseActorEnrichmentMeta'
      - type: object
...
    ActorEnrichmentMeta:
      type: object
      description: This is the unified interface for actor enrichment algorithm meta
        informaiton.
      discriminator:
        propertyName: enrichment_type
        mapping:
          category: '#/components/schemas/CategoryActorEnrichmentMeta'
          numerical: '#/components/schemas/NumericalActorEnrichmentMeta'
          array: '#/components/schemas/ArrayActorEnrichmentMeta'
          text: '#/components/schemas/TextActorEnrichmentMeta'
      oneOf:
      - $ref: '#/components/schemas/CategoryActorEnrichmentMeta'
      - $ref: '#/components/schemas/NumericalActorEnrichmentMeta'
      - $ref: '#/components/schemas/ArrayActorEnrichmentMeta'
      - $ref: '#/components/schemas/TextActorEnrichmentMeta'
...
    BaseActorEnrichment:
      required:
      - enrichmentType
      type: object
      properties:
        actorUuid:
          type: string
        providerName:
          type: string
          description: "The team (e.g., USC) who provides the enrichment."
        enrichmentName:
          type: string
          description: "The enrichment (e.g., age, gender) name for the enrichment."
        version:
          type: string
          description: "The version within the same (provider, enrichmentName)."
        confidence:
          maximum: 1.0
          minimum: 0.0
          type: number
          description: "The confidence that this enrichment is correct, expressed\
            \ as a percentage between 0.0 and 1.0"
          format: double
        enrichmentType:
          type: string
          description: For discriminator
          enum:
          - category
          - numerical
          - array
          - text
      description: Stores data fields common to all types of actor enrichment.
      discriminator:
        propertyName: enrichmentType
        mapping:
          category: '#/components/schemas/CategoryActorEnrichment'
          numerical: '#/components/schemas/NumericalActorEnrichment'
          array: '#/components/schemas/ArrayActorEnrichment'
          text: '#/components/schemas/TextActorEnrichment'
    CategoryActorEnrichment:
      description: Defines a enumeration/category enrichment for actor.
      allOf:
      - $ref: '#/components/schemas/BaseActorEnrichment'
      - type: object
        properties:
          attributeValue:
            type: string
    NumericalActorEnrichment:
      description: Defines a numerical enrichment for actor.
      allOf:
      - $ref: '#/components/schemas/BaseActorEnrichment'
      - type: object
        properties:
          attributeValue:
            type: number
    ArrayActorEnrichment:
      description: Defines a vector/matrix/tensor or multi-label enrichment for actor.
      allOf:
      - $ref: '#/components/schemas/BaseActorEnrichment'
      - type: object
        properties:
          attributeValue:
            type: array
            items:
              type: number
              format: float
    TextActorEnrichment:
      description: This defines a text enrichment for actor.
      allOf:
      - $ref: '#/components/schemas/BaseActorEnrichment'
      - type: object
        properties:
          attributeValue:
            type: string
    ActorEnrichment:
      type: object
      description: This is the unified interface for actor enrichment output.
      discriminator:
        propertyName: enrichment_type
        mapping:
          enumeration: '#/components/schemas/CategoryActorEnrichment'
          numerical: '#/components/schemas/NumericalActorEnrichment'
          array: '#/components/schemas/ArrayActorEnrichment'
          text: '#/components/schemas/TextActorEnrichment'
      oneOf:
      - $ref: '#/components/schemas/CategoryActorEnrichment'
      - $ref: '#/components/schemas/NumericalActorEnrichment'
      - $ref: '#/components/schemas/ArrayActorEnrichment'
      - $ref: '#/components/schemas/TextActorEnrichment'
...
    Actor:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        actorName:
          type: string
        links:
          $ref: '#/components/schemas/Links'
        entityType:
          type: string
          enum:
          - ENTITY_UNSPECIFIED
          - Person
          - Media
          - Organization
          - Government
        mediaResources:
          type: array
          items:
            $ref: '#/components/schemas/MediaResource'
        knowledgeBaseUrl:
          type: string
        language:
          type: string
        description:
          type: string
        annotations:
          type: array
          items:
            $ref: '#/components/schemas/Annotation'
        extraAttributes:
          type: array
          items:
            $ref: '#/components/schemas/ExtraAttribute'
      example:
        knowledgeBaseUrl: knowledgeBaseUrl
        mediaResources:
        - accountStatus: accountStatus
          accountId: accountId
          hashedUserNames:
          - hashedUserNames
          - hashedUserNames
          accountType: MEDIA_UNSPECIFIED
          displayNames:
          - displayNames
          - displayNames
          verified: true
          userNames:
          - userNames
          - userNames
          url: url
        - accountStatus: accountStatus
          accountId: accountId
          hashedUserNames:
          - hashedUserNames
          - hashedUserNames
          accountType: MEDIA_UNSPECIFIED
          displayNames:
          - displayNames
          - displayNames
          verified: true
          userNames:
          - userNames
          - userNames
          url: url
        entityType: ENTITY_UNSPECIFIED
        name: name
        description: description
        annotations:
        - offsets:
          - end: 6
            begin: 0
          - end: 6
            begin: 0
          confidence: 1.4658129805029452
          name: name
          id: id
          text: text
          type: type
          providerName: providerName
        - offsets:
          - end: 6
            begin: 0
          - end: 6
            begin: 0
          confidence: 1.4658129805029452
          name: name
          id: id
          text: text
          type: type
          providerName: providerName
        actorName: actorName
        links:
          self: self
        language: language
        id: id
        extraAttributes:
        - attributeValue: attributeValue
          name: name
          id: id
          attributeKey: attributeKey
          providerName: providerName
        - attributeValue: attributeValue
          name: name
          id: id
          attributeKey: attributeKey
          providerName: providerName
    Links:
      type: object
      properties:
        self:
          type: string
      example:
        self: self
    MediaResource:
      type: object
      properties:
        accountType:
          type: string
          enum:
          - MEDIA_UNSPECIFIED
          - News
          - Reddit
          - Twitter
          - Webpage
        accountId:
          type: string
        userNames:
          type: array
          items:
            type: string
        hashedUserNames:
          type: array
          items:
            type: string
        accountStatus:
          type: string
        displayNames:
          type: array
          items:
            type: string
        verified:
          type: boolean
        url:
          type: string
      example:
        accountStatus: accountStatus
        accountId: accountId
        hashedUserNames:
        - hashedUserNames
        - hashedUserNames
        accountType: MEDIA_UNSPECIFIED
        displayNames:
        - displayNames
        - displayNames
        verified: true
        userNames:
        - userNames
        - userNames
        url: url
    Annotation:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        providerName:
          type: string
        type:
          type: string
        offsets:
          type: array
          items:
            $ref: '#/components/schemas/Offset'
        text:
          type: string
        confidence:
          type: number
          format: double
      example:
        offsets:
        - end: 6
          begin: 0
        - end: 6
          begin: 0
        confidence: 1.4658129805029452
        name: name
        id: id
        text: text
        type: type
        providerName: providerName
    Offset:
      type: object
      properties:
        begin:
          type: integer
          format: int64
        end:
          type: integer
          format: int64
      example:
        end: 6
        begin: 0
    ExtraAttribute:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        providerName:
          type: string
        attributeKey:
          type: string
        attributeValue:
          type: string
      example:
        attributeValue: attributeValue
        name: name
        id: id
        attributeKey: attributeKey
        providerName: providerName
...